+++ core/arch/arm/plat-sam/main.c
@@ -41,7 +41,6 @@
 #include <mm/core_memprot.h>
 #include <platform_config.h>
 #include <sama5d2.h>
-#include <sam_sfr.h>
 #include <stdint.h>
 #include <sm/optee_smc.h>
 #include <tz_matrix.h>
@@ -54,20 +53,6 @@ void console_init(void)
 {
 	atmel_uart_init(&console_data, CONSOLE_UART_BASE);
 	register_serial_console(&console_data.chip);
-}
-
-register_phys_mem_pgdir(MEM_AREA_IO_SEC, SFR_BASE, CORE_MMU_PGDIR_SIZE);
-
-vaddr_t sam_sfr_base(void)
-{
-	static void *va;
-
-	if (cpu_mmu_enabled()) {
-		if (!va)
-			va = phys_to_virt(SFR_BASE, MEM_AREA_IO_SEC, 1);
-		return (vaddr_t)va;
-	}
-	return SFR_BASE;
 }
 
 register_phys_mem_pgdir(MEM_AREA_IO_SEC, AT91C_BASE_MATRIX32,
